Extract event publishers to persistence-model
