+++ persistence-model/src/main/scala/hmda/persistence/model/HmdaPersistentActor.scala
@@ -1,4 +1,4 @@
-package hmda.persistence
+package hmda.persistence.model
 
 import java.util.concurrent.TimeUnit
 
@@ -7,7 +7,6 @@ import akka.persistence.PersistentActor
 import akka.stream.ActorMaterializer
 import com.typesafe.config.ConfigFactory
 import hmda.persistence.messages.CommonMessages.{ Event, Shutdown }
-import hmda.persistence.model.HmdaActor
 
 import scala.concurrent.duration.Duration
