+++ persistence-model/src/main/scala/hmda/persistence/model/LocalEventPublisher.scala
@@ -1,4 +1,5 @@
-package hmda.persistence
+package hmda.persistence.model
+
 import akka.actor.ActorSystem
 import hmda.persistence.messages.CommonMessages.Event
