+++ persistence/src/main/scala/hmda/persistence/processing/HmdaFileParser.scala
@@ -7,7 +7,7 @@ import hmda.model.fi.ts.TransmittalSheet
 import hmda.parser.fi.lar.LarCsvParser
 import hmda.parser.fi.ts.TsCsvParser
 import hmda.persistence.messages.CommonMessages._
-import hmda.persistence.{ HmdaPersistentActor, LocalEventPublisher }
+import hmda.persistence.model.{ HmdaPersistentActor, LocalEventPublisher }
 import hmda.persistence.processing.HmdaQuery._
 import hmda.persistence.processing.HmdaRawFile.LineAdded
