+++ persistence/src/main/scala/hmda/persistence/processing/HmdaRawFile.scala
@@ -3,7 +3,7 @@ package hmda.persistence.processing
 import akka.actor.{ ActorRef, ActorSystem, Props }
 import hmda.model.fi.SubmissionId
 import hmda.persistence.messages.CommonMessages._
-import hmda.persistence.{ HmdaPersistentActor, LocalEventPublisher }
+import hmda.persistence.model.{ HmdaPersistentActor, LocalEventPublisher }
 
 object HmdaRawFile {
